backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
Authlib==1.6.0
boto3==1.40.19
botocore==1.40.19
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
deprecation==2.1.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.13
frozenlist==1.7.0
GeoAlchemy2==0.18.0
gotrue==2.12.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-health-checking==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflect==7.5.0
iniconfig==2.1.0
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
more-itertools==10.7.0
multidict==6.5.0
packaging==25.0
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.2
protobuf==6.31.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
realtime==2.4.3
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
sqlacodegen==3.1.0
sqlacodegen-v2==0.1.4
SQLAlchemy==2.0.41
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.3
supafunc==0.9.4
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
validators==0.35.0
weaviate-client==4.16.1
websockets==14.2
yarl==1.20.1
